requirements.txt
asgiref==3.7.2
astroid==2.15.8
attrs==23.2.0
colorama==0.4.6
Django==3.2
django-dotenv==1.4.2
djangorestframework==3.14.0
drf-spectacular==0.27.1
drf-yasg==1.21.7
Faker==12.0.1
flake8==6.1.0
flake8-django==1.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.10.0
mccabe==0.7.0
mixer==7.2.2
packaging==23.2
pluggy==1.4.0
pycodestyle==2.11.1
pyflakes==3.1.0
pytest==8.0.0
pytest-django==4.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.17.1
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
wrapt==1.16.0
